+++ src/main/java/main/repositories/Tag2PostRepository.java
@@ -15,5 +15,5 @@ public interface Tag2PostRepository extends JpaRepository<Tag2Post, Long> {
      * @return - возвращает количество связок пост-тэг конкретного поста.
      */
     @Query("SELECT tp FROM Tag2Post tp WHERE tp.post.id = :postId ORDER BY tp.tag.id")
-    List<Tag2Post> findAllTag2PostByPostId(@Param("postId") int postId);
+    List<Tag2Post> findAllTag2PostByPostId(@Param("postId") long postId);
 }
