+++ src/main/java/main/responses/UserWithPhotoDTO.java
@@ -8,7 +8,7 @@ import lombok.NoArgsConstructor;
 public class UserWithPhotoDTO extends UserSimpleDTO {
     private String photo;
 
-    public UserWithPhotoDTO(int id, String name, String photo) {
+    public UserWithPhotoDTO(long id, String name, String photo) {
         super(id, name);
         this.photo = photo;
     }
