+++ src/main/java/main/services/PostVoteService.java
@@ -1,14 +1,14 @@
 package main.services;
 
 public interface PostVoteService {
-    int getCountLikesByPostId(int postId);
-    int getCountDislikesByPostId(int postId);
+    int getCountLikesByPostId(long postId);
+    int getCountDislikesByPostId(long postId);
     int getTotalCountLikes();
     int getTotalCountDislikes();
-    int getTotalCountLikesByUserId(int userId);
-    int getTotalCountDislikesByUserId(int userId);
-    boolean userLikeAlreadyExists(int userId, int postId);
-    boolean userDislikeAlreadyExists(int userId, int postId);
-    int getIdByUserIdAndPostId(int userId, int postId);
-    void deleteById(int postVoteId);
+    int getTotalCountLikesByUserId(long userId);
+    int getTotalCountDislikesByUserId(long userId);
+    boolean userLikeAlreadyExists(long userId, long postId);
+    boolean userDislikeAlreadyExists(long userId, long postId);
+    int getIdByUserIdAndPostId(long userId, long postId);
+    void deleteById(long postVoteId);
 }
